Allow a zoom of 0 in the URL
This zoom level is possible by zooming out via control buttons or mouse
wheel, so it should also be possible by passing via URL instead of
falling back to 10.

After this bugfix, passing an empty zoom value (z=) in the URL now
causes the exception resulting in the location with default coordinates
and zoom. The fallback zoom of 10 is not used anymore, but was
inconsistent with the default zoom of 13 in src/leaflet_options.js
anyway.