Update fr.js
Start, end and via strings are too long and repetitive (it's always " ... - add a placename and press Enter") so I would keep it simple in the text boxes and add text in an onHover pop up to give the user more explanations.  It could also be mentioned that once a route in on the map arbitrary waypoints can be added by simply dragging the road with the mouse.