Fix link parsing for waypoints
Previously, waypoint parsing from links only worked iff more than one
`loc' was given in the link because qs parses several locs to an array
but only one loc to a string. Therefore we check if `q.loc' is an array
or a string and adapt the waypoint parsing accordingly.