Added service selector dropdown.
When the 'services' property contains a list of more than one
service, a drop-down menu will appear in the bottom left of the
map allowing the user to switch the router between any of the
configured services.

It is also now possible to specify a 'profile' for a service so
that the default of 'driving' can be changed. This 'profile'
affects the URL used when connecting to the service, e.g.:

  http://127.0.0.1:5000/route/v1/<profile>/13.3888,52.5170;...