Default to english translation
This introduces a new function t(), which can be called to retrieve a
translation for a certain key value in a certain language.
If no translation is found in the current language, it defaults to
english. If no english translation is found, it returns the given key.
This is to avoid displaying "undefined" or no message at all.