Move config to config.json file
This commit allows the user to have greater control of the services,
layers and overlays in the frontend. Instead of only being able to send
one backend variable, which would then be replaced into a javscript file
read as a text file by hardcoded values, this file can be passed to the
docker container with the -v command.

The README file has been updated to describe the changes.